Disable queuing of indexing-operations during migration to Dexterity to prevent catalog-errors.
