+++ CHANGES.rst
@@ -34,6 +34,9 @@ Bug fixes:
   definitions for the content types that have extra field defined on top of the
   behavior composition. Otherwise no translations can be picked up.
   [fredvd]
+- Disable queuing of indexing-operations (PLIP https://github.com/plone/Products.CMFPlone/issues/1343)
+  during migration to Dexterity to prevent catalog-errors.
+  [pbauer]
 
 - Use original raw text and mimetype when indexing rich text.
   This avoids a double transform (raw source to output mimetype to plain text).
